# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "The RESTFul API of Server",
        "title": "ACG.Fate API",
        "contact": {},
        "version": "0.0.1"
    },
    "host": "127.0.0.1:3000",
    "basePath": "/api/v1",
    "paths": {
        "/check/username": {
            "get": {
                "description": "查询用户名是否存在",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Check"
                ],
                "summary": "查询用户名是否存在",
                "parameters": [
                    {
                        "type": "string",
                        "name": "username",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "0": {
                        "description": "msg: \"可用",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    },
                    "40021": {
                        "description": "msg: 用户名已存在",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    }
                }
            }
        },
        "/email/verify": {
            "get": {
                "description": "验证码邮件发送接口",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Email"
                ],
                "summary": "验证码邮件发送",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "0": {
                        "description": "msg: \"Success",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    },
                    "30001": {
                        "description": "msg: 参数错误",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "验证邮件发送接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Email"
                ],
                "summary": "验证邮件发送",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "验证码",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.MailVerifyService"
                        }
                    }
                ],
                "responses": {
                    "0": {
                        "description": "msg: \"Success",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    },
                    "30001": {
                        "description": "msg: 参数错误",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    }
                }
            }
        },
        "/sign": {
            "post": {
                "description": "普通签到接口",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Task"
                ],
                "summary": "普通签到",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "0": {
                        "description": "msg: \"Success",
                        "schema": {
                            "$ref": "#/definitions/serializer.TaskSignResponse"
                        }
                    },
                    "60001": {
                        "description": "msg: 参数错误",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    }
                }
            }
        },
        "/user/email": {
            "get": {
                "description": "查询邮箱是否存在",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Check"
                ],
                "summary": "查询邮箱是否存在",
                "parameters": [
                    {
                        "type": "string",
                        "name": "email",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "0": {
                        "description": "msg: \"可用",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    },
                    "40022": {
                        "description": "msg: 邮箱已存在",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    }
                }
            }
        },
        "/user/info": {
            "get": {
                "description": "用户个人信息接口",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "用户个人信息",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "0": {
                        "description": "msg: Success",
                        "schema": {
                            "$ref": "#/definitions/serializer.BasicInfoResponse"
                        }
                    },
                    "50000": {
                        "description": "msg: 查询个人信息错误",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    }
                }
            }
        },
        "/user/login": {
            "post": {
                "description": "用户登录接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "用户登录",
                "parameters": [
                    {
                        "description": "用户名, 密码",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.LoginService"
                        }
                    }
                ],
                "responses": {
                    "0": {
                        "description": "msg: Success",
                        "schema": {
                            "$ref": "#/definitions/serializer.LoginResponse"
                        }
                    },
                    "30001": {
                        "description": "msg: 参数错误",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    }
                }
            }
        },
        "/user/register": {
            "post": {
                "description": "用户注册接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "用户注册",
                "parameters": [
                    {
                        "description": "用户名, 密码, 邮箱, 昵称",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.RegisterService"
                        }
                    }
                ],
                "responses": {
                    "0": {
                        "description": "msg: Success",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    },
                    "30001": {
                        "description": "msg: 参数错误",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    }
                }
            }
        },
        "/user/update": {
            "put": {
                "description": "用户信息更新接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "用户信息更新",
                "parameters": [
                    {
                        "type": "string",
                        "description": "用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "用户信息",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.UpdateInfoService"
                        }
                    }
                ],
                "responses": {
                    "0": {
                        "description": "msg: \"Success",
                        "schema": {
                            "$ref": "#/definitions/serializer.BasicInfoResponse"
                        }
                    },
                    "30001": {
                        "description": "msg: 参数错误",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    }
                }
            }
        },
        "/words/:id": {
            "get": {
                "description": "文字查看接口",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Words"
                ],
                "summary": "文字查看",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Words ID",
                        "name": "WID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "0": {
                        "description": "msg:\"获取成功",
                        "schema": {
                            "$ref": "#/definitions/serializer.WordsResponse"
                        }
                    },
                    "30001": {
                        "description": "msg: 参数错误",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    }
                }
            }
        },
        "/words/post": {
            "post": {
                "description": "文字发表接口",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Words"
                ],
                "summary": "文字发表",
                "parameters": [
                    {
                        "description": "文字发表信息",
                        "name": "form",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/services.PostService"
                        }
                    }
                ],
                "responses": {
                    "0": {
                        "description": "msg:\"发表成功",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    },
                    "30001": {
                        "description": "msg: 参数错误",
                        "schema": {
                            "$ref": "#/definitions/serializer.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "serializer.BasicInfoResponse": {
            "type": "object",
            "properties": {
                "birthday": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "credit": {
                    "type": "integer"
                },
                "exp": {
                    "type": "integer"
                },
                "gender": {
                    "type": "string"
                },
                "join_time": {
                    "type": "string"
                },
                "level": {
                    "type": "integer"
                },
                "nickname": {
                    "type": "string"
                },
                "province": {
                    "type": "string"
                },
                "sign": {
                    "type": "string"
                },
                "uid": {
                    "type": "integer"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "serializer.LoginResponse": {
            "type": "object",
            "properties": {
                "token": {
                    "description": "Token",
                    "type": "string"
                },
                "uid": {
                    "description": "UID",
                    "type": "integer"
                }
            }
        },
        "serializer.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "object"
                },
                "data": {
                    "type": "object"
                },
                "msg": {
                    "type": "string"
                }
            }
        },
        "serializer.TaskSignResponse": {
            "type": "object",
            "properties": {
                "reward": {
                    "type": "string"
                }
            }
        },
        "serializer.WordsResponse": {
            "type": "object",
            "properties": {
                "author": {
                    "type": "string"
                },
                "category": {
                    "type": "string"
                },
                "comments": {
                    "type": "integer"
                },
                "content": {
                    "type": "string"
                },
                "edit_time": {
                    "type": "string"
                },
                "likes": {
                    "type": "integer"
                },
                "post_time": {
                    "type": "string"
                },
                "tags": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "views": {
                    "type": "integer"
                },
                "wid": {
                    "type": "integer"
                }
            }
        },
        "services.LoginService": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "services.MailVerifyService": {
            "type": "object",
            "required": [
                "code"
            ],
            "properties": {
                "code": {
                    "type": "string"
                }
            }
        },
        "services.PostService": {
            "type": "object",
            "required": [
                "category",
                "content"
            ],
            "properties": {
                "category": {
                    "type": "integer"
                },
                "content": {
                    "type": "string"
                },
                "tags": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "services.RegisterService": {
            "type": "object",
            "required": [
                "email",
                "nickname",
                "password",
                "username"
            ],
            "properties": {
                "email": {
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        },
        "services.UpdateInfoService": {
            "type": "object",
            "properties": {
                "birthday": {
                    "type": "string"
                },
                "city": {
                    "type": "string"
                },
                "gender": {
                    "description": "只能为 男、女、其他、保密",
                    "type": "string"
                },
                "nickname": {
                    "type": "string"
                },
                "province": {
                    "type": "string"
                },
                "sign": {
                    "type": "string"
                }
            }
        }
    }
}